requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-solo==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.5
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-telegram-bot==20.2
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
